[ui] Made QueueInfo into editable component
