+++ packages/ui/src/components/dashboard/analytics/Layout.tsx
@@ -1,21 +1,41 @@
-import { Grid, GridItem } from "@chakra-ui/react";
+import {
+  Grid,
+  GridItem,
+  Button,
+  useBoolean,
+  Box,
+  VStack,
+} from "@chakra-ui/react";
 import QR from "./QR";
 
 import BanWidget from "./widgets/BanWidget";
 import QueueInfoWidget from "./widgets/QueueInfoWidget";
 import QueueListWidget from "./widgets/QueueListWidget";
+import { BsFillPenFill } from "react-icons/bs";
 
 const Layout = ({ queuedUsers, queueInfo }) => {
+  const [canEdit, toggleCanEdit] = useBoolean(false);
+
   return (
     <Grid
       pl="1%"
       pt="1%"
       grid-template-rows="auto auto "
       grid-template-columns="auto auto "
-      gap="5%"
+      gap="10%"
     >
       <GridItem colStart={1} colEnd={3} rowStart={1} rowEnd={2}>
-        <QueueInfoWidget queueInfo={queueInfo}></QueueInfoWidget>
+        <Box>
+          <VStack>
+            <Button
+              onClick={toggleCanEdit.toggle}
+              rightIcon={<BsFillPenFill />}
+            >
+              Edit Queue
+            </Button>
+            <QueueInfoWidget queueInfo={queueInfo} canEdit={canEdit} />
+          </VStack>
+        </Box>
       </GridItem>
       <GridItem colStart={1} colEnd={3} rowStart={2} rowEnd={3}>
         <QueueListWidget queuedUsers={queuedUsers}></QueueListWidget>
